requirement.txt
langchain==0.3.0
langchain_community
langchain_google_genai 
google-generativeai 
pandas 
streamlit 
requests
scikit-learn
python-dotenv
Flask